core/requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==4.2.17
django-debug-toolbar==5.0.1
django-filter==24.3
django-mail-templated==2.6.5
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
